Refactor edit-user-profile
