+++ .expo/types/router.d.ts
@@ -7,7 +7,7 @@ declare module "expo-router" {
   export * from 'expo-router/build';
 
   // prettier-ignore
-  type StaticRoutes = `/` | `/(onboarding)/_layout` | `/_layout` | `/(onboarding)/address` | `/address` | `/(onboarding)/bio` | `/bio` | `/(onboarding)/jobs` | `/jobs` | `/(onboarding)/location` | `/location` | `/(onboarding)/onboarding` | `/onboarding` | `/(onboarding)/provider-location` | `/provider-location` | `/(onboarding)/provider-type` | `/provider-type` | `/(onboarding)/shops` | `/shops` | `/(onboarding)/welcome-provider` | `/welcome-provider` | `/(onboarding)/welcome` | `/welcome` | `/(register)/_layout` | `/(register)/attempt` | `/attempt` | `/(register)/login` | `/login` | `/(register)/password` | `/password` | `/(register)/prepare` | `/prepare` | `/(register)/register` | `/register` | `/(register)/success` | `/success` | `/(tabs)/(profile)/_layout` | `/(tabs)/(profile)/change-password` | `/change-password` | `/(tabs)/(profile)/edit-user-profile` | `/edit-user-profile` | `/(tabs)/(profile)/profile` | `/profile` | `/(tabs)/_layout` | `/(tabs)/home` | `/home`;
+  type StaticRoutes = `/` | `/(onboarding)/_layout` | `/_layout` | `/(onboarding)/address` | `/address` | `/(onboarding)/bio` | `/bio` | `/(onboarding)/jobs` | `/jobs` | `/(onboarding)/location` | `/location` | `/(onboarding)/onboarding` | `/onboarding` | `/(onboarding)/provider-location` | `/provider-location` | `/(onboarding)/provider-type` | `/provider-type` | `/(onboarding)/shops` | `/shops` | `/(onboarding)/welcome-provider` | `/welcome-provider` | `/(onboarding)/welcome` | `/welcome` | `/(register)/_layout` | `/(register)/attempt` | `/attempt` | `/(register)/login` | `/login` | `/(register)/password` | `/password` | `/(register)/prepare` | `/prepare` | `/(register)/register` | `/register` | `/(register)/success` | `/success` | `/(tabs)/(profile)/_layout` | `/(tabs)/(profile)/change-password` | `/change-password` | `/(tabs)/(profile)/edit-user-profile` | `/edit-user-profile` | `/(tabs)/_layout` | `/(tabs)/home` | `/home` | `/(tabs)/(profile)/user-profile` | `/user-profile`;
   // prettier-ignore
   type DynamicRoutes<T extends string> = never;
   // prettier-ignore
