+++ src/app/(tabs)/(profile)/edit-user-profile.tsx
@@ -1,40 +1,29 @@
 import { useUser } from "@clerk/clerk-expo";
-import { Ionicons } from "@expo/vector-icons";
-import BottomSheet, { BottomSheetBackdrop, BottomSheetBackdropProps } from "@gorhom/bottom-sheet";
+import BottomSheet from "@gorhom/bottom-sheet";
 import { zodResolver } from "@hookform/resolvers/zod";
-import * as FileSystem from "expo-file-system";
 import * as ImagePicker from "expo-image-picker";
 import { Link, Stack, useRouter } from "expo-router";
-import React, { useCallback, useMemo, useRef } from "react";
+import React, { useRef } from "react";
 import { useForm } from "react-hook-form";
 import {
   StyleSheet,
   View,
-  Image,
   TextInput,
   Text,
   TouchableOpacity,
   Button,
   Alert,
+  Keyboard,
 } from "react-native";
 import Spinner from "react-native-loading-spinner-overlay";
 import Toast from "react-native-root-toast";
 import * as z from "zod";
 
+import ImagePickerBottomSheet from "@/components/bottom/ImagePickerBottomSheet";
 import ControlledInput from "@/components/inputs/ControlledInput";
-import Colors from "@/constants/Colors";
+import ProfileImage from "@/components/profile/ProfileImage";
 import { defaultStyles } from "@/constants/Styles";
-
-// [x] Cambiar foto
-// [x] Guardar cambios
-// [x] Cambiar contraseña
-// [x] Labels
-// [x] Guardar
-// [x] Contraseña
-// [x] Confirmar eliminar foto
-// [X] Toast perfil actualizado
-// [x] Spinner
-// [] Limpiar
+import { processImage } from "@/utils/image";
 
 type FormData = {
   firstName: string;
@@ -57,8 +46,8 @@ const schema = z.object({
 const EditUserProfileScreen = () => {
   const { user } = useUser();
   const router = useRouter();
+
   const bottomSheetRef = useRef<BottomSheet>(null);
-  const snapPoints = useMemo(() => ["32%"], []);
   const areaCode = user?.phoneNumbers![0].phoneNumber.slice(0, 3);
 
   const {
@@ -73,20 +62,6 @@ const EditUserProfileScreen = () => {
     },
     mode: "onChange",
   });
-
-  const processImage = async (uri: string) => {
-    try {
-      const base64Image = await FileSystem.readAsStringAsync(uri, {
-        encoding: FileSystem.EncodingType.Base64,
-      });
-
-      const base64Prefix = "data:image/png;base64,";
-      return base64Image.startsWith(base64Prefix) ? base64Image : base64Prefix + base64Image;
-    } catch (error) {
-      console.error("Error processing image: ", error);
-      alert("Error processing the image.");
-    }
-  };
 
   const launchGallery = async () => {
     try {
@@ -162,15 +137,9 @@ const EditUserProfileScreen = () => {
     );
   };
 
-  const renderBackdrop = useCallback(
-    (props: BottomSheetBackdropProps) => (
-      <BottomSheetBackdrop {...props} disappearsOnIndex={-1} appearsOnIndex={1} />
-    ),
-    []
-  );
-
   // Function to handle saving changes
   const saveChanges = async (data: FormData) => {
+    Keyboard.dismiss();
     try {
       await user?.update({
         firstName: data.firstName,
@@ -207,19 +176,7 @@ const EditUserProfileScreen = () => {
         }}
       />
       <View style={[defaultStyles.container, { alignItems: "center", gap: 20 }]}>
-        <View style={{ position: "relative" }}>
-          <Image style={styles.profileImage} source={{ uri: user?.imageUrl }} />
-          <View style={styles.editIconContainer}>
-            <Ionicons
-              name="pencil"
-              size={18}
-              color="grey"
-              onPress={() => {
-                bottomSheetRef.current?.expand();
-              }}
-            />
-          </View>
-        </View>
+        <ProfileImage user={user} bottomSheetRef={bottomSheetRef} />
         <View
           style={{
             gap: 5,
@@ -254,8 +211,6 @@ const EditUserProfileScreen = () => {
           style={{
             gap: 5,
             width: "100%",
-            // alignItems: "flex-start",
-            // marginHorizontal: "auto",
           }}
         >
           <Text>E-mail</Text>
@@ -271,8 +226,6 @@ const EditUserProfileScreen = () => {
             style={{
               gap: 5,
               width: "100%",
-              // alignItems: "flex-start",
-              // marginHorizontal: "auto",
             }}
           >
             <Text>Contraseña</Text>
@@ -296,9 +249,6 @@ const EditUserProfileScreen = () => {
         <View
           style={{
             gap: 5,
-            // width: "100%",
-            // alignItems: "flex-start",
-            // marginHorizontal: "auto",
           }}
         >
           <Text>Teléfono</Text>
@@ -321,47 +271,13 @@ const EditUserProfileScreen = () => {
             />
           </View>
         </View>
-        <BottomSheet
+        <ImagePickerBottomSheet
           ref={bottomSheetRef}
-          index={-1}
-          snapPoints={snapPoints}
-          backdropComponent={renderBackdrop}
-          enablePanDownToClose
-        >
-          <View
-            style={{
-              alignItems: "flex-start",
-              gap: 20,
-              marginVertical: 20,
-              paddingHorizontal: 16,
-            }}
-          >
-            <Text style={{ fontFamily: "mon-sb", fontSize: 16 }}>Editar foto de perfil </Text>
-            <TouchableOpacity
-              onPress={launchCamera}
-              style={{ flexDirection: "row", gap: 10, alignItems: "center" }}
-            >
-              <Ionicons name="camera" size={24} color="black" />
-              <Text>Tomar foto</Text>
-            </TouchableOpacity>
-            <TouchableOpacity
-              onPress={launchGallery}
-              style={{ flexDirection: "row", gap: 10, alignItems: "center" }}
-            >
-              <Ionicons name="image" size={24} color="black" />
-              <Text>Seleccionar de la galeria</Text>
-            </TouchableOpacity>
-
-            <TouchableOpacity
-              onPress={deletePhoto}
-              style={{ flexDirection: "row", gap: 10, alignItems: "center" }}
-              disabled={!user?.hasImage}
-            >
-              <Ionicons name="trash" size={24} color={!user?.hasImage ? "grey" : Colors.error} />
-              <Text style={{ color: !user?.hasImage ? "grey" : Colors.error }}>Eliminar foto</Text>
-            </TouchableOpacity>
-          </View>
-        </BottomSheet>
+          user={user}
+          launchCamera={launchCamera}
+          launchGallery={launchGallery}
+          deletePhoto={deletePhoto}
+        />
       </View>
     </>
   );
@@ -370,30 +286,6 @@ const EditUserProfileScreen = () => {
 export default EditUserProfileScreen;
 
 const styles = StyleSheet.create({
-  profileImage: {
-    width: 100,
-    height: 100,
-    borderRadius: 50,
-  },
-  editIconContainer: {
-    position: "absolute",
-    bottom: 0,
-    right: 0,
-    backgroundColor: "#fdffff",
-    padding: 5,
-    width: 30,
-    height: 30,
-    borderRadius: 5,
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 5,
-  },
   input: {
     height: 40,
     width: "100%",
